minor changes + add README
